refactor: extract scroll-into-view logic into RowScroller collaborator
